initial commit
This would be the initial working version.   So far the idea is to define an extractor interface that all feature extraction objects follow. Each  Feature extraction object is responsible for extracting  individual features alongside corresponding statistics.  The pipeline object simply  contains the image and applies the  extract  function of each extractor object.